# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Polygon Demo",
        "description": "A NodeJs application to polygon demo",
        "license": {
            "name": "MIT"
        }
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/polygon/create-did": {
            "post": {
                "tags": [
                    "Create polygon DID"
                ],
                "summary": "API Endpoint for create polygon DID",
                "parameters": [{
                    "name": "body",
                    "in": "body",
                    "description": "Create polygon DID",
                    "required": true,
                    "type": "string",
                    "schema": {
                        "properties": {
                            "networkType": {
                                "type": "string",
                                "example": "testnet"
                            },
                            "privateKey": {
                                "type": "string",
                                "example": "0xd922b5e24ab1132f68d8c3a3e758d6814f71d5b1acffdffa6a19b8f8d55db507"
                            }
                        }
                    }
                }],
                "responses": {
                    "201": {
                        "description": "New Create polygon DID"
                    }
                }
            }
        },
        "/polygon/register-did": {
            "post": {
                "tags": [
                    "Register polygon DID"
                ],
                "summary": "API Endpoint for register polygon DID",
                "description": "Register polygon DID",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Register new polygon DID",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "properties": {
                            "did": {
                                "type": "string",
                                "example": "did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993"
                            },
                            "privateKey": {
                                "type": "string",
                                "example": "0xd922b5e24ab1132f68d8c3a3e758d6814f71d5b1acffdffa6a19b8f8d55db507"
                            }
                        }
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "New register polygon DID"
                    }
                }
            }
        },
        "/polygon/resolve-did": {
            "post": {
                "tags": [
                    "Resolve polygon DID"
                ],
                "summary": "API Endpoint for resolve polygon DID",
                "description": "Resolve polygon DID",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Resolve new polygon DID",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "properties": {
                            "did": {
                                "type": "string",
                                "example": "did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993"
                            }
                        }
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "New resolve polygon DID"
                    }
                }
            }
        },
        "/polygon/update-did": {
            "post": {
                "tags": [
                    "Update polygon DID"
                ],
                "summary": "API Endpoint for update polygon DID",
                "description": "Update polygon DID",
                "parameters": [{
                    "name": "body",
                    "in": "body",
                    "description": "Update new polygon DID",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "properties": {
                            "did": {
                                "type": "string",
                                "example": "did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993"
                            },
                            "didDoc": {
                                "type": "object",
                                "example": {
                                    "@context": "https://w3id.org/did/v1",
                                    "id": "did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993",
                                    "verificationMethod": [{
                                        "id": "did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993",
                                        "type": "EcdsaSecp256k1VerificationKey2019",
                                        "controller": ["did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993", "did:polygon:0x794b781493AeD65b9ceBD680716fec257e118993"],
                                        "publicKeyBase58": "7Lnm1ZnseKDkH1baAb1opREfAU4MPY7zCdUDSrWSm9NxNTQmy4neU9brFUYnEcyy7CwFKjD11ikyP9J8cf6zEaAKrEzzp"
                                    }]
                                }
                            },
                            "privateKey": {
                                "type": "string",
                                "example": "0xd922b5e24ab1132f68d8c3a3e758d6814f71d5b1acffdffa6a19b8f8d55db507"
                            }
                        }
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "New update polygon DID"
                    }
                }
            }
        },
        "/polygon/delete-did": {
            "post": {
                "tags": [
                    "Delete polygon DID"
                ],
                "summary": "API Endpoint for delete polygon DID",
                "description": "Update polygon DID",
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Delete polygon DID",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "properties": {
                            "did": {
                                "type": "string",
                                "example": "did:polygon:testnet:0x794b781493AeD65b9ceBD680716fec257e118993"
                            },
                            "privateKey": {
                                "type": "string",
                                "example": "0xd922b5e24ab1132f68d8c3a3e758d6814f71d5b1acffdffa6a19b8f8d55db507"
                            }
                        }
                    }
                }],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Delete polygon DID"
                    }
                }
            }
        }
    }
}